+++ website/src/content/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Added
+
+- Added `Table` component.
+
 ## [0.13.0] - 2023-09-19
 
 ### Added